Websocket improvements
Based on the changes to the server side, on clientside, a socket is opening only when a student click a topic cell. Further, if a client tries to open another websocket connection within 10 seconds, it is being penalized and the connection request did not send to the server. This is to eliminate an unknown error of infinite socket openning which causes the apache become unresponsive.